--- test/utils/test-setup.ts
@@ -1,7 +1,0 @@
-import "source-map-support/register";
-import "reflect-metadata";
-import * as chai from "chai";
-
-chai.should();
-chai.use(require("sinon-chai"));
-chai.use(require("chai-as-promised"));
